extract out info panel
