+++ src/style/clone-table.module.css
@@ -19,10 +19,12 @@
 
 .container .row :global(.ant-table-cell) {
     border-bottom: none;
-    padding: 10px;
+    padding: 5px 16px;
+    line-height: 1.4;
 }
 
 .footer {
-    margin: 4px 0 20px 0;
+    margin: 4px 0 8px 0;
     font-size: 12px;
+    line-height: 1.2;
 }
