+++ src/style/info-panel.module.css
@@ -1,0 +1,26 @@
+.container :global(.ant-descriptions) {
+    border-radius: 4px;
+    border: 1px var(--small-table-border-color) solid;
+    margin-bottom: 20px;
+}
+
+.container :global(.ant-descriptions-view) table {
+    border-collapse: collapse;
+}
+
+.container :global(.ant-descriptions-item) {
+    border-spacing: 0;
+    border-bottom: 1px var(--small-table-border-color) solid;
+    padding: 10px 0 10px 20px !important;
+    line-height: 1;
+}
+
+.container
+    :global(.ant-descriptions-row):last-of-type
+    :global(.ant-descriptions-item) {
+    border-bottom: none;
+}
+
+.container tbody {
+    border-spacing: 10px;
+}
