+++ src/style/main-card.module.css
@@ -1,8 +1,13 @@
+.container {
+    border: 1px solid var(--DEEP_BLUE);
+    border-radius: 4px;
+}
+
 .container :global(.ant-card-head) {
     background-color: var(--main-card-header-color);
     color: white;
     padding: 16px 28px;
-    border-radius: 4px 4px 0 0;
+    border-radius: 3px 3px 0 0;
     height: 79px;
 }
 
@@ -10,6 +15,10 @@
     font-size: 48px;
     font-weight: 600;
     color: white;
+}
+
+.container :global(.ant-card-body) {
+    padding: 16px 20px;
 }
 
 .container :global(.ant-card-head-title .ant-btn) {
@@ -24,28 +33,38 @@
     border: 1px solid white;
 }
 
-.container :global(.ant-descriptions) {
-    border-radius: 4px;
-    border: 1px var(--small-table-border-color) solid;
-    margin-bottom: 20px;
+.container .small-button {
+    width: 100%;
+    height: 40px;
+    border: 1px solid var(--DEEP_BLUE);
+    margin-top: 8px;
 }
 
-.container :global(.ant-descriptions-view) table {
-    border-collapse: collapse;
+.container .small-button span {
+    line-height: 30px;
 }
 
-.container :global(.ant-descriptions-item) {
-    border-spacing: 0;
-    border-bottom: 1px var(--small-table-border-color) solid;
-    padding: 14.5px 0 14.5px 20px !important;
+.container .small-button:hover {
+    background-color: var(--OFF_WHITE) !important;
 }
 
-.container
-    :global(.ant-descriptions-row):last-of-type
-    :global(.ant-descriptions-item) {
-    border-bottom: none;
+.extra-large-button {
+    width: 100%;
+    height: 102px;
+    border: 2.5px solid var(--DEEP_BLUE);
+    margin-top: 20px;
 }
 
-.container tbody {
-    border-spacing: 10px;
+.extra-large-button h2 {
+    font-size: 24px;
+    font-weight: 600;
+}
+
+.extra-large-button:hover {
+    transition: background 0.2s cubic-bezier(0.645, 0.045, 0.355, 1);
+    background: linear-gradient(
+        180deg,
+        #0b9aab 0%,
+        rgba(11, 154, 171, 0.5) 100%
+    ) !important;
 }
